Implement reducers as class methods
