+++ ui/actions.py
@@ -1,24 +1,109 @@
+import os
+import pydux
+from pydux.extend import extend
 import utility
+from functools import partial
+from bookfile_list import BookFile_List
+import pprint
 
-class Reducers(object):
-    def set_books(state):
-        pass
-    def go_to_book(state):
-        pass
-    def go_to_library(state):
-        pass
-    def go_to_menu(state):
-        pass
-    def next_page(state):
-        pass
-    def previous_page(state):
-        pass
-    def replace_library(state):
-        pass
-    def shutdown(state):
-        pass
-    def backup_log(state):
-        pass
+def dimensions(state):
+    width = state['display']['width']
+    height = state['display']['height']
+    return [width, height]
+
+
+def get_title(book):
+    return utility.alphas_to_pin_nums(os.path.basename(book['data'].filename))
+
+
+def set_page(book, page, height):
+    data = book['data']
+    if page < 0 or page > get_max_pages(data, height):
+        return book
+    else:
+        return {'data': data, 'page': page}
+
+
+def get_max_pages(data, height):
+    return len(data) // height
+
+
+class Reducers():
+    @staticmethod
+    def go_to_book(state, action):
+        width, height = dimensions(state)
+        page = state['library']['page']
+        line_number = page * height
+        try:
+            location = state['books'][line_number + action['value']]
+        except:
+            log.warning('no book at {}'.format(action['value']))
+            return state
+        return extend(state, {'location': line_number + action['value']})
+    @staticmethod
+    def go_to_library(state, action):
+        return extend(state, {'location': 'library'})
+    @staticmethod
+    def go_to_menu(state, action):
+        return extend(state, {'location': 'menu'})
+    @staticmethod
+    def set_books(state, action):
+        print(state)
+        width, height = dimensions(state)
+        books = []
+        for filename in action['value']:
+            books.append(
+                {
+                    'data': BookFile_List(filename, [width, height]),
+                    'page': 0
+                }
+            )
+        data = map(get_title, books)
+        data = map(partial(utility.pad_line, width), data)
+        library = {'data': data, 'page': 0}
+        return extend(state, {'books': tuple(books), 'library': library})
+    @staticmethod
+    def next_page(state, action):
+        width, height = dimensions(state)
+        location = state['location']
+        if location == 'library':
+            library = state['library']
+            page    = state['library']['page'] + 1
+            library = set_page(library, page, height)
+            return extend(state, {'library': library})
+        elif type(location) == int:
+            book = state['books'][location]
+            data = book['data']
+            page = book['page'] + 1
+            books = list(state['books'])
+            books[location] = set_page(book, page, height)
+            return extend(state, {'books': tuple(books)})
+    @staticmethod
+    def previous_page(state, action):
+        width, height = dimensions(state)
+        location = state['location']
+        if location == 'library':
+            library = state['library']
+            page    = library['page'] - 1
+            library = set_page(library, page, height)
+            return extend(state, {'library': library})
+        elif type(location) == int:
+            book = state['books'][location]
+            data = book['data']
+            page = book['page'] - 1
+            books = list(state['books'])
+            books[location] = set_page(book, page, height)
+            return extend(state, {'books': tuple(books)})
+    @staticmethod
+    def replace_library(state, action):
+        return state
+    @staticmethod
+    def shutdown(state, action):
+        return state
+    @staticmethod
+    def backup_log(state, action):
+        return state
+
 
 action_types = utility.get_methods(Reducers)
 
@@ -33,3 +118,4 @@ def actions(): pass
 #then we give it a method for each action
 for action in action_types:
     setattr(actions, action, make_action_method(action))
+
