+++ ui/utility.py
@@ -10,7 +10,6 @@ import re
 import logging
 import functools
 log = logging.getLogger(__name__)
-from types import FunctionType
 
 class FormfeedConversionException(Exception): pass
 class LinefeedConversionException(Exception): pass
@@ -162,5 +161,6 @@ def pad_line(w, line):
     line.extend([0] * (w - len(line)))
     return line
 
-def get_methods (cls):
-    return [x for x,y in cls.__dict__.items() if type(y) == FunctionType]
+def get_methods(cls):
+    methods = [x for x in dir(cls) if callable(getattr(cls, x))]
+    return filter(lambda x: not x.startswith('__'), methods)
